Commit API Token and a part of User

+++ src/main/java/Server/controller/CommentController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.CommentDAO;
 import Server.model.DB.CommentEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,43 +16,57 @@ import java.util.List;
 public class CommentController {
     @Autowired
     CommentDAO commentDAO;
-    @RequestMapping(value = "/Post",
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
+    @RequestMapping(value = "{apiToken}/Post",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> postCategorySong(@RequestBody CommentEntity entity){
+    public ResponseEntity<?> postCategorySong(@PathVariable("apiToken") String apiToken, @RequestBody CommentEntity entity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         commentDAO.save(entity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
-    @RequestMapping(value = "Put/{id}",
+
+    @RequestMapping(value = "{apiToken}/Put/{id}",
             method = RequestMethod.PUT)
     @ResponseBody
-    public  ResponseEntity<?> updateCategorySong (@RequestBody CommentEntity entity, @PathVariable("id") Long id){
-        if(entity.getId()==id)
-        {
-            commentDAO.save(entity);
-            return new ResponseEntity<>("Update Completed",HttpStatus.OK);
+    public ResponseEntity<?> updateCategorySong(@PathVariable("apiToken") String apiToken, @RequestBody CommentEntity entity, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return new ResponseEntity<>("Update Fail",HttpStatus.BAD_REQUEST);
+        if (entity.getId() == id) {
+            commentDAO.save(entity);
+            return new ResponseEntity<>("Update Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Update Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value = "/Delete/{id}",
+
+    @RequestMapping(value = "{apiToken}/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> delete(@PathVariable("id") Long id){
-        if(commentDAO.getByID(id)!=null){
+    public ResponseEntity<?> delete(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
+        if (commentDAO.getByID(id) != null) {
             commentDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
-        }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value="/GetComment/model={model}&id={id}",
-        method = RequestMethod.GET,  produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/GetComment/model={model}&id={id}",
+            method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public ResponseEntity<?> getComment(@PathVariable("id") Long id,@PathVariable("model") String model){
-        List<CommentEntity> commentEntityList = commentDAO.getId(model,id);
-        if(!commentEntityList.isEmpty()){
-            return new ResponseEntity<>(commentEntityList,HttpStatus.OK);
+    public ResponseEntity<?> getComment(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id, @PathVariable("model") String model) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        return new ResponseEntity<>("Please create first comment",HttpStatus.OK);
+        List<CommentEntity> commentEntityList = commentDAO.getId(model, id);
+        if (!commentEntityList.isEmpty()) {
+            return new ResponseEntity<>(commentEntityList, HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Please create first comment", HttpStatus.OK);
     }
 }
