+++ src/main/java/Server/controller/FilmController.java
@@ -21,60 +21,76 @@ public class FilmController {
     FilmDAO filmDAO;
     @Autowired
     FilmSiteDAO filmSiteDAO;
-    @RequestMapping(value = "/Post",
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
+    @RequestMapping(value = "{apiToken}/Post",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> post(@RequestBody FilmEntity entity){
-        entity=filmDAO.save(entity);
+    public ResponseEntity<?> post(@PathVariable("apiToken") String apiToken, @RequestBody FilmEntity entity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
+        entity = filmDAO.save(entity);
 
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
-    @RequestMapping(value = "/Put/{id}",
+
+    @RequestMapping(value = "{apiToken}/Put/{id}",
             method = RequestMethod.PUT)
     @ResponseBody
-    public  ResponseEntity<?> update (@RequestBody FilmEntity entity, @PathVariable("id") Long id){
-        if(filmDAO.getByID(id)!=null)
-        {
-            filmDAO.save(entity);
-            return new ResponseEntity<>("Update Completed",HttpStatus.OK);
+    public ResponseEntity<?> update(@PathVariable("apiToken") String apiToken, @RequestBody FilmEntity entity, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return new ResponseEntity<>("Update Fail",HttpStatus.BAD_REQUEST);
+        if (filmDAO.getByID(id) != null) {
+            filmDAO.save(entity);
+            return new ResponseEntity<>("Update Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Update Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value = "/Delete/{id}",
+
+    @RequestMapping(value = "{apiToken}/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> delete(@PathVariable("id") Long id){
-        if(filmDAO.getByID(id)!=null){
-            filmDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
+    public ResponseEntity<?> delete(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        if (filmDAO.getByID(id) != null) {
+            filmDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value = "/GetDetail/{id}",
-            method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
-    @ResponseBody
-    public  ResponseEntity<?> getDetail (@PathVariable("id") Long id){
 
-        return new ResponseEntity<>(filmSiteDAO.getFilmDTOById(id),HttpStatus.OK);
-    }
-    @RequestMapping(value = "/GetTop{item}" , method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+    @RequestMapping(value = "{apiToken}/GetDetail/{id}",
+            method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public ResponseEntity<?> getTop10(@PathVariable int item)
-    {
-        try{
+    public ResponseEntity<?> getDetail(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(filmSiteDAO.getFilmDTOById(id), HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "{apiToken}/GetTop{item}", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
+    @ResponseBody
+    public ResponseEntity<?> getTop10(@PathVariable("apiToken") String apiToken, @PathVariable int item) {
+        try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             Criteria criteria = new Criteria();
             criteria.setClazz(FilmEntity.class);
             criteria.setTop(item);
             List<FilmEntity> lsFilm = filmDAO.GetTop10(criteria);
             List<FilmDTO> lsFilmDTO = new ArrayList<>();
-            for ( FilmEntity filmE : lsFilm
-                 ) {
+            for (FilmEntity filmE : lsFilm
+            ) {
                 lsFilmDTO.add(filmSiteDAO.getFilmDTOById(filmE.getId()));
             }
-            return new ResponseEntity<>(lsFilmDTO,HttpStatus.OK);
+            return new ResponseEntity<>(lsFilmDTO, HttpStatus.OK);
         } catch (Exception e) {
             LogEntity log = new LogEntity(e);
             (new LogDAO()).save(log);
@@ -82,64 +98,74 @@ public class FilmController {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
-    @RequestMapping(value ="/GetAllHasPage{itemOnPage}/{page}", method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/GetAllHasPage{itemOnPage}/{page}", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public  ResponseEntity<?> getPage (@PathVariable("page") int page,@PathVariable("itemOnPage") int itemOnPage){
-        try{
+    public ResponseEntity<?> getPage(@PathVariable("apiToken") String apiToken, @PathVariable("page") int page, @PathVariable("itemOnPage") int itemOnPage) {
+        try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             Criteria criteria = new Criteria();
             criteria.setClazz(FilmEntity.class);
             criteria.setCurrentPage(page);
             criteria.setItemPerPage(itemOnPage);
             List<FilmEntity> lsFilm = filmDAO.loadDataPagination(criteria);
             List<FilmDTO> lsFilmDTO = new ArrayList<>();
-            for ( FilmEntity filmE : lsFilm)
-            {
+            for (FilmEntity filmE : lsFilm) {
                 lsFilmDTO.add(filmSiteDAO.getFilmDTOById(filmE.getId()));
             }
-            return new ResponseEntity<>(lsFilmDTO,HttpStatus.OK);
-       }catch (Exception e) {
+            return new ResponseEntity<>(lsFilmDTO, HttpStatus.OK);
+        } catch (Exception e) {
             LogEntity log = new LogEntity(e);
             (new LogDAO()).save(log);
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
-    @RequestMapping(value ="/Count", method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/Count", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public  ResponseEntity<?> count (){
+    public ResponseEntity<?> count(@PathVariable("apiToken") String apiToken) {
         try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             return new ResponseEntity<>(filmDAO.count(), HttpStatus.OK);
         } catch (Exception e) {
             new LogDAO().save(new LogEntity(e));
             e.printStackTrace();
-            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg", HttpStatus.BAD_REQUEST);
         }
     }
-    @RequestMapping(value ="/GetRandom{item}",
+
+    @RequestMapping(value = "{apiToken}/GetRandom{item}",
             method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public  ResponseEntity<?> getRandom (@PathVariable("item") int item) {
+    public ResponseEntity<?> getRandom(@PathVariable("apiToken") String apiToken, @PathVariable("item") int item) {
         try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             Criteria criteria = new Criteria();
             criteria.setClazz(FilmEntity.class);
             criteria.setTop(item);
             List<FilmEntity> filmEntityList = filmDAO.loadTopRandom(criteria);
             List<FilmDTO> lsFilmDTO = new ArrayList<>();
-            for ( FilmEntity filmE : filmEntityList)
-            {
+            for (FilmEntity filmE : filmEntityList) {
                 lsFilmDTO.add(filmSiteDAO.getFilmDTOById(filmE.getId()));
             }
-            if(!filmEntityList.isEmpty())
+            if (!filmEntityList.isEmpty())
                 return new ResponseEntity<>(lsFilmDTO, HttpStatus.OK);
-            return new ResponseEntity<>("Not Found",HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             LogEntity log = new LogEntity(e);
             (new LogDAO()).save(log);
             e.printStackTrace();
-            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg", HttpStatus.BAD_REQUEST);
         }
     }
 }
