+++ src/main/java/Server/controller/ImageController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.ImageDAO;
 import Server.model.DB.ImageEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,33 +13,42 @@ import org.springframework.web.bind.annotation.*;
 public class ImageController {
     @Autowired
     ImageDAO imageDAO;
-
-    @RequestMapping(value = "/Post",
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+    @RequestMapping(value = "{apiToken}/Post",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> post(@RequestBody ImageEntity entity){
+    public ResponseEntity<?> post(@PathVariable("apiToken") String apiToken, @RequestBody ImageEntity entity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         imageDAO.save(entity);
-        return new ResponseEntity<>("Post Completed",HttpStatus.CREATED);
+        return new ResponseEntity<>("Post Completed", HttpStatus.CREATED);
     }
-    @RequestMapping(value = "/Put/{id}",
+
+    @RequestMapping(value = "{apiToken}/Put/{id}",
             method = RequestMethod.PUT)
     @ResponseBody
-    public  ResponseEntity<?> update(@RequestBody ImageEntity entity, @PathVariable("id") Long id){
-        if(id==entity.getId())
-        {imageDAO.save(entity);
-        return new ResponseEntity<>("Update completed",HttpStatus.OK);
+    public ResponseEntity<?> update(@PathVariable("apiToken") String apiToken, @RequestBody ImageEntity entity, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return new ResponseEntity<>("Update Fail",HttpStatus.BAD_REQUEST);
+        if (id == entity.getId()) {
+            imageDAO.save(entity);
+            return new ResponseEntity<>("Update completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Update Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value = "/Delete/{id}",
+
+    @RequestMapping(value = "{apiToken}/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> delete(@PathVariable("id") Long id){
-        if(imageDAO.getByID(id)!=null){
-            imageDAO.delete(id);
-            return new ResponseEntity<>("Deleted completed",HttpStatus.OK);
+    public ResponseEntity<?> delete(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return new ResponseEntity<>("Deleted Fail",HttpStatus.BAD_REQUEST);
+        if (imageDAO.getByID(id) != null) {
+            imageDAO.delete(id);
+            return new ResponseEntity<>("Deleted completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Deleted Fail", HttpStatus.BAD_REQUEST);
     }
 }
