+++ src/main/java/Server/controller/SeriFilmController.java
@@ -24,65 +24,82 @@ public class SeriFilmController {
     SeriFilmDAO seriFilmDAO;
     FilmSiteDAO filmSiteDAO = new FilmSiteDAO();
     SerifilmFilmDAO serifilmFilmDAO = new SerifilmFilmDAO();
-    @RequestMapping(value = "/Post",
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
+    @RequestMapping(value = "{apiToken}/Post",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> post(@RequestBody SerifilmEntity entity ){
+    public ResponseEntity<?> post(@PathVariable("apiToken") String apiToken, @RequestBody SerifilmEntity entity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         seriFilmDAO.save(entity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
-    @RequestMapping(value = "/Put/{id}",
+
+    @RequestMapping(value = "{apiToken}/Put/{id}",
             method = RequestMethod.PUT)
     @ResponseBody
-    public  ResponseEntity<?> updateAccount(@RequestBody SerifilmEntity entity, @PathVariable("id") Long id){
-        if(seriFilmDAO.getByID(id)!=null)
-        {
-            seriFilmDAO.save(entity);
-            return new ResponseEntity<>("Update Completed",HttpStatus.OK);
+    public ResponseEntity<?> updateAccount(@PathVariable("apiToken") String apiToken, @RequestBody SerifilmEntity entity, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return new ResponseEntity<>("Update Fail",HttpStatus.BAD_REQUEST);
+        if (seriFilmDAO.getByID(id) != null) {
+            seriFilmDAO.save(entity);
+            return new ResponseEntity<>("Update Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Update Fail", HttpStatus.BAD_REQUEST);
     }
 
-    @RequestMapping(value = "/Delete/{id}",
+    @RequestMapping(value = "{apiToken}/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteAccount(@PathVariable("id") Long id){
-        if(seriFilmDAO.getByID(id)!=null){
-            seriFilmDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
+    public ResponseEntity<?> deleteAccount(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        if (seriFilmDAO.getByID(id) != null) {
+            seriFilmDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
-    @RequestMapping(value = "/GetDetail/{id}",
+
+    @RequestMapping(value = "{apiToken}/GetDetail/{id}",
             method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE}
+            produces = {MediaType.APPLICATION_JSON_VALUE}
     )
     @ResponseBody
-    public ResponseEntity<?> getDetail(@PathVariable("id") Long id){
-        SerifilmEntity serifilmEntity = seriFilmDAO.getByID(id);
-        if(serifilmEntity!=null){
-            return  new ResponseEntity<>(filmSiteDAO.getSeriFilmDTO(serifilmEntity),HttpStatus.OK);
+    public ResponseEntity<?> getDetail(@PathVariable("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        return new ResponseEntity<>("Not Found",HttpStatus.NOT_FOUND);
+        SerifilmEntity serifilmEntity = seriFilmDAO.getByID(id);
+        if (serifilmEntity != null) {
+            return new ResponseEntity<>(filmSiteDAO.getSeriFilmDTO(serifilmEntity), HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
     }
-    @RequestMapping(value = "/GetTop{iTop}" , method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/GetTop{iTop}", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public ResponseEntity<?> getTop10(@PathVariable("iTop") int iTop) {
+    public ResponseEntity<?> getTop10(@PathVariable("apiToken") String apiToken, @PathVariable("iTop") int iTop) {
         try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             Criteria criteria = new Criteria();
             criteria.setClazz(SerifilmEntity.class);
             criteria.setTop(iTop);
-            List<SerifilmEntity> serifilmEntityList =seriFilmDAO.getTop10(criteria);
+            List<SerifilmEntity> serifilmEntityList = seriFilmDAO.getTop10(criteria);
             List<SeriFilmDTO> seriFilmDTOList = new ArrayList<>();
-            if(!serifilmEntityList.isEmpty()){
-                for(SerifilmEntity item : serifilmEntityList){
+            if (!serifilmEntityList.isEmpty()) {
+                for (SerifilmEntity item : serifilmEntityList) {
                     seriFilmDTOList.add(filmSiteDAO.getSeriFilmDTO(item));
                 }
                 return new ResponseEntity<>(seriFilmDTOList, HttpStatus.OK);
             }
-            return new ResponseEntity<>("Not Found",HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             LogEntity log = new LogEntity(e);
             (new LogDAO()).save(log);
@@ -90,41 +107,49 @@ public class SeriFilmController {
             return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg", HttpStatus.BAD_REQUEST);
         }
     }
-    @RequestMapping(value ="/GetAllHasPage{itemOnPage}/{page}", method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/GetAllHasPage{itemOnPage}/{page}", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public  ResponseEntity<?> getPage (@PathVariable("page") int page,@PathVariable("itemOnPage") int itemOnPage ){
-        try{
-        Criteria criteria = new Criteria();
-        criteria.setClazz(SerifilmEntity.class);
-        criteria.setCurrentPage(page);
-        criteria.setItemPerPage(itemOnPage);
-        List<SerifilmEntity> serifilmEntityList = seriFilmDAO.loadDataPagination(criteria);
-        List<SeriFilmDTO> seriFilmDTOList = new ArrayList<>();
-        if(!seriFilmDTOList.isEmpty()){
-            for(SerifilmEntity item: serifilmEntityList){
-                seriFilmDTOList.add(filmSiteDAO.getSeriFilmDTO(item));
+    public ResponseEntity<?> getPage(@PathVariable("apiToken") String apiToken, @PathVariable("page") int page, @PathVariable("itemOnPage") int itemOnPage) {
+        try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
             }
-            return new ResponseEntity<>(seriFilmDTOList,HttpStatus.OK);
-        }
-        return new ResponseEntity<>("Not Found",HttpStatus.NOT_FOUND);
+            Criteria criteria = new Criteria();
+            criteria.setClazz(SerifilmEntity.class);
+            criteria.setCurrentPage(page);
+            criteria.setItemPerPage(itemOnPage);
+            List<SerifilmEntity> serifilmEntityList = seriFilmDAO.loadDataPagination(criteria);
+            List<SeriFilmDTO> seriFilmDTOList = new ArrayList<>();
+            if (!seriFilmDTOList.isEmpty()) {
+                for (SerifilmEntity item : serifilmEntityList) {
+                    seriFilmDTOList.add(filmSiteDAO.getSeriFilmDTO(item));
+                }
+                return new ResponseEntity<>(seriFilmDTOList, HttpStatus.OK);
+            }
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             LogEntity log = new LogEntity(e);
             (new LogDAO()).save(log);
             e.printStackTrace();
-            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg", HttpStatus.BAD_REQUEST);
         }
     }
-    @RequestMapping(value ="/Count", method = RequestMethod.GET,
-            produces = { MediaType.APPLICATION_JSON_VALUE})
+
+    @RequestMapping(value = "{apiToken}/Count", method = RequestMethod.GET,
+            produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
-    public  ResponseEntity<?> count (){
+    public ResponseEntity<?> count(@PathVariable("apiToken") String apiToken) {
         try {
+            if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+                return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+            }
             return new ResponseEntity<>(seriFilmDAO.count(), HttpStatus.OK);
         } catch (Exception e) {
             new LogDAO().save(new LogEntity(e));
             e.printStackTrace();
-            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>("If you are admin, Check table Log to see ErrorMsg", HttpStatus.BAD_REQUEST);
         }
     }
 }
